refactor: extract parseWorkspaceActions + prompt to protocols/WorkspaceActions module
Move WorkspaceAction interface, WORKSPACE_ACTIONS_PROMPT constant,
VALID_ACTION_TYPES set, and parseWorkspaceActions function from
AppServer.ts to src/services/protocols/WorkspaceActions.ts. Update
imports in AppServer.ts, SessionManager.ts, and test file.
